fix(Plurals): use decimal values

+++ src/Plurals/RuleContext.cs
@@ -21,7 +21,7 @@ namespace Sepia.Globalization.Plurals
         /// <summary>
         ///   Integer digits of n.
         /// </summary>
-        public int i;
+        public decimal i;
 
         /// <summary>
         ///   Number of visible fraction digits in n, with trailing zeros.
@@ -59,7 +59,7 @@ namespace Sepia.Globalization.Plurals
             var context = new RuleContext
             {
                 n = Math.Abs(value),
-                i = (int) Math.Floor(value),
+                i = Math.Floor(value),
             };
             if (dot >= 0)
             {
